requirements.txt
streamlit
PyPDF2
python-docx
spacy
nltk
sumy

